dev_requirements.txt
-r requirements.txt

black==18.9b0
pre-commit==1.14.4
django-extensions==2.1.6
